Make the command interface a little more robust
Added three optional values minArgs, maxArgs and description.
minArgs and maxArgs are used to validate arg count before calling a
command and displaying help if the number of arguments aren't in the
range.
description is displayed with over all help to help describe a command.

This also converts help from a method to just a string since that makes
more sense.