Add a module to track the modes of a nick in a channel
This adds a mode_tracker which keeps track of what modes a nick has in
the given channel. It also has a way of querying to see if a nick has at
least given level of permission.

When queried for a mode check if the module doesn't have the nick's
modes it will use whois to determine op or voice.